chore: move index and browser entry points to src
